# Rewrite binary search implementation