@Amanieu 2024-07-26
Benchmark results:

### Old

```
test binary_search_l1              ... bench:          28.09 ns/iter (+/- 0.74)
test binary_search_l1_with_dups    ... bench:          18.12 ns/iter (+/- 0.35)
test binary_search_l1_worst_case   ... bench:          14.19 ns/iter (+/- 1.14)
test binary_search_l2              ... bench:          40.88 ns/iter (+/- 1.11)
test binary_search_l2_with_dups    ... bench:          27.79 ns/iter (+/- 0.70)
test binary_search_l2_worst_case   ... bench:          25.57 ns/iter (+/- 0.83)
test binary_search_l3              ... bench:         109.16 ns/iter (+/- 1.25)
test binary_search_l3_with_dups    ... bench:          70.63 ns/iter (+/- 4.93)
test binary_search_l3_worst_case   ... bench:          44.09 ns/iter (+/- 0.74)
```

### New

```
test binary_search_l1              ... bench:          11.87 ns/iter (+/- 0.98)
test binary_search_l1_with_dups    ... bench:          11.73 ns/iter (+/- 0.47)
test binary_search_l1_worst_case   ... bench:           9.80 ns/iter (+/- 0.22)
test binary_search_l2              ... bench:          18.23 ns/iter (+/- 3.47)
test binary_search_l2_with_dups    ... bench:          17.57 ns/iter (+/- 1.12)
test binary_search_l2_worst_case   ... bench:          14.70 ns/iter (+/- 1.17)
test binary_search_l3              ... bench:          61.18 ns/iter (+/- 1.63)
test binary_search_l3_with_dups    ... bench:          61.67 ns/iter (+/- 2.21)
test binary_search_l3_worst_case   ... bench:          24.73 ns/iter (+/- 0.64)
```