@tgross35 2024-07-27
The result change for equal values seems fine to me, considering docs explicitly allow it. We could also consider special-casing on `T::IS_ZST` either to ensure those results do not change, or to provide an algorithm-agnostic deterministic output like #127007.

I think this is easily justified by the performance wins. However, since it does change user-visible stable behavior, it seems reasonable to get group consensus or decide if crater results are wanted first (as suggested at https://github.com/rust-lang/rust/blob/21ee15741d020022d3b6736b378240dd1a697743/library/core/tests/slice.rs#L83-L97).

@rustbot label +I-libs-nominated

r=me for the actual implementation change after the above is addressed in some way, and a rebase once https://github.com/rust-lang/rust/pull/128250 merges.