@Amanieu 2024-07-28
I don't think it is worth special casing ZSTs since realistically nobody is ever going to sort a ZST slice. As long as we're still generating correct results for ZSTs, I would just ignore them.