@joboet 2024-07-29
I'd like to mention @scandum's [collection of binary search algorithms](https://github.com/scandum/binary_search), the current version in this PR is identical to the "monobound binary search". But maybe it'd be worth switching to the "tripple-tapped" variant instead?