@Amanieu 2024-07-29
The problem with the triple-tapped variant is that the number of iterations in the second loop is data-dependent. One of the main reasons the current algorithm is so fast is that it achieves perfect branch prediction no matter which key is being searched for.