@Amanieu 2024-07-30
This crater run shouldn't block merging this PR, it's just there so that in a later beta crater run we can better identify failures caused by the binary search implementation change.

@craterbot test