@tgross35 2024-07-30
This isn't blocked on anything per https://hackmd.io/zSmxqkrmTE21keL4HgAKKg#nominated-rusttf128254-Rewrite-binary-search-implementation. Perfbot is almost free so may as well see the results before merge.

@rust-timer queue