@tgross35 2024-07-30
That doesn't seem right, there shouldn't need to be a rebuild.

@rust-timer build 5e89635d7bf1d3033c2eb20378b7b3cbe2dd1a9b