@Amanieu 2024-07-30
ucd basically [entirely consists of calls to `binary_search_by`](https://github.com/rust-lang/rustc-perf/blob/master/collector/compile-benchmarks/ucd/src/tables/mod.rs). Most of the regression is just LLVM spending more time optimizing the binary search implementation. Considering the benchmark results, I think this is acceptable.