@Kobzol 2024-08-06
This is an optimization of the binary search algorithm, which should make runtime faster. The regression seems to be caused by LLVM spending more time in optimizing the binary search (since it is now more amenable to optimization and unrolling).

@rustbot label: +perf-regression-triaged