@Amanieu 2024-08-09
@rust-lang/release

Refer to these crater results when evaluating the beta crater run: these are crates whose tests depends on implementation details of `binary_search` when there are multiple matching elements. This PR changed the exact index that is returned in such cases, which is allowed by the documentation of `binary_search`.